Factor out MonitorInspectorTable
