+++ src/screens/home/admin/MonitorInspectorLampsCard.tsx
@@ -1,16 +1,7 @@
 import { TitledCard } from '@luna/components/TitledCard';
 import { LampV2Metrics } from '@luna/contexts/api/model/types';
-import { MonitorInspectorValue } from '@luna/screens/home/admin/MonitorInspectorValue';
-import {
-  Table,
-  TableCell,
-  TableColumn,
-  TableHeader,
-  TableRow,
-} from '@nextui-org/react';
+import { MonitorInspectorTable } from '@luna/screens/home/admin/MonitorInspectorTable';
 import { IconLamp } from '@tabler/icons-react';
-import { useMemo } from 'react';
-import { TableBody } from 'react-stately';
 
 export interface MonitorInspectorLampsCardProps {
   metrics: LampV2Metrics[];
@@ -29,44 +20,9 @@ const names: { [Property in keyof LampV2Metrics]: string } = {
 export function MonitorInspectorLampsCard({
   metrics,
 }: MonitorInspectorLampsCardProps) {
-  const columns = [...Array(metrics.length + 1).keys()].map(i => ({
-    key: i,
-  }));
-
-  const rows = useMemo(
-    () =>
-      metrics.length > 0
-        ? (Object.keys(metrics[0]) as (keyof LampV2Metrics)[]).map(
-            (prop, i) => ({
-              key: i,
-              lampValues: [names[prop], ...metrics.map(lamp => lamp[prop])],
-            })
-          )
-        : [],
-    [metrics]
-  );
-
   return (
     <TitledCard icon={<IconLamp />} title="Lamps">
-      <Table hideHeader isStriped isCompact aria-label="Lamp monitoring values">
-        <TableHeader columns={columns}>
-          {column => <TableColumn key={column.key}>{column.key}</TableColumn>}
-        </TableHeader>
-        <TableBody items={rows}>
-          {item => (
-            <TableRow key={item.key}>
-              {columnKey => {
-                const i = columnKey as number;
-                return (
-                  <TableCell>
-                    <MonitorInspectorValue value={item.lampValues[i]} />
-                  </TableCell>
-                );
-              }}
-            </TableRow>
-          )}
-        </TableBody>
-      </Table>
+      <MonitorInspectorTable metrics={metrics} names={names} />
       {/* <div className="flex flex-row">
         {metrics
           ? metrics.map((lamp: any, idx: number) => (
